+++ be/src/main/java/com/be/auth/kakao/dto/KakaoAccount.java
@@ -1,4 +1,4 @@
-package com.be.auth.dto.kakao;
+package com.be.auth.kakao.dto;
 
 public record KakaoAccount(
         String email,
